refactor: organize main layer folders
