--- src/main/factories/controllers/login/login-controller-factory.ts
@@ -1,10 +1,0 @@
-import { LoginController } from '../../../../presentation/controllers/login/login/login-controller'
-import { Controller } from '../../../../presentation/protocols'
-import { makeLoginValidation } from './login-validation-factory'
-import { makeDbAuthentication } from '../../usecases/authentication/db-authentication-factory'
-import { makeLogControllerDecorator } from '../../decorators/log-controller-decorator-factory'
-
-export const makeLoginController = (): Controller => {
-  const controller = new LoginController(makeDbAuthentication(), makeLoginValidation())
-  return makeLogControllerDecorator(controller)
-}
