+++ src/main/factories/controllers/login/signup/signup-controller-factory.ts
@@ -1,0 +1,11 @@
+import { SignUpController } from '../../../../../presentation/controllers/login/signup/signup-controller'
+import { Controller } from '../../../../../presentation/protocols'
+import { makeSignUpValidation } from './signup-validation-factory'
+import { makeDbAuthentication } from '../../../usecases/account/authentication/db-authentication-factory'
+import { makeDbAddAccount } from '../../../usecases/account/add-account/db-add-account-factory'
+import { makeLogControllerDecorator } from '../../../decorators/log-controller-decorator-factory'
+
+export const makeSignUpController = (): Controller => {
+  const controller = new SignUpController(makeDbAddAccount(), makeSignUpValidation(), makeDbAuthentication())
+  return makeLogControllerDecorator(controller)
+}
