--- src/main/factories/usecases/add-account/db-add-account-factory.ts
@@ -1,11 +1,0 @@
-import { BcryptAdapter } from '../../../../infra/criptography/bcrypt/bcrypt-adapter'
-import { AccountMongoRepository } from '../../../../infra/db/mongodb/account/account-mongo-repository'
-import { DbAddAccount } from '../../../../data/usecases/add-account/db-add-account'
-import { AddAccount } from '../../../../data/usecases/add-account/db-add-account-protocols'
-
-export const makeDbAddAccount = (): AddAccount => {
-  const salt = 12
-  const bcryptAdapter = new BcryptAdapter(salt)
-  const accountMongoRepository = new AccountMongoRepository()
-  return new DbAddAccount(bcryptAdapter, accountMongoRepository, accountMongoRepository)
-}
